Alterações diversas
Cadeira:  Incluindo a configuração para horário de aulas
(ConfiguracaoCadeira) e alteração para lista de Planejamentos (1 por
período)
Falta: simplificação do serviço para inclusão de faltas (falta alterar
para toggle)
Aula: Inclusao do Periodo na Aula
Avaliacao: inclusao da data da avaliacao
Educando: troca da ordem de byte para int

+++ iSchool.Api.Core/Model/Educando.cs
@@ -9,7 +9,7 @@ namespace iSchool.Api.Core.Model
 	public class Educando : IEntityBase
 	{
 		public int Id { get; set; }
-		public byte OrdemChamada { get; set; }
+		public int OrdemChamada { get; set; }
 		public int alunoId { get; set; }
 		public Aluno Aluno { get; set; }
 		public List<Cadeira> Aulas { get; set; }
